### 2021-11-22 #3944 @whedon
[PRE REVIEW]: Rainbow: Automated Air-Liquid Interface Cell Culture Analysis Using Deep Optical Flow
**Submitting author:** @AlphonsG (<a href="http://orcid.org/0000-0002-9607-2839">Alphons Gwatimba</a>)
**Repository:** <a href="https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI" target ="_blank">https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI</a>
**Version:** 0.1.0
