### 2021-11-22 #3944 @Kevin-Mattheus-Moerman
@AlphonsG I am using the correct branch there, i.e. `dev` to generate the paper from?
