+++ paper/paper.md
@@ -14,6 +14,7 @@ authors:
     orcid: 0000-0002-9607-2839
     affiliation: "1, 2"
   - name: Joseph Ho
+    orcid: 0000-0002-9496-506X
     affiliation: "1, 2"
   - name: Thomas Iosifidis^[co-second author]
     orcid: 0000-0001-8462-5865
@@ -39,11 +40,11 @@ bibliography: paper.bib
 
 # Summary
 
-`Rainbow` is a free, open-source and cross-platform Python-based tool for Air-Liquid Interface (ALI) cell culture image analysis. `Rainbow` quantifies cell motion in image sequences using a recent state-of-the-art deep learning based optical flow model  [@jiangLearningEstimateHidden2021]. Optical flow is the apparent motion of individual pixels on an image plane across frames [@turagaAdvancesVideoBasedHuman2010]. Our software generates motion magnitude and direction variance-based metrics to provide quantitative insight into heterogeneous cell dynamics observed in the ALI culture images, in addition to commonly used metrics such as average cell speed. We also include a unified hue/saturation-based visualisation scheme for cell motion which offers increased resolution as every coordinate on an image is utilised. `Rainbow` performs data analysis detailing cell dynamics by generating a HTML report containing output images, videos, publication ready figures, and p-values from input images provided in standard image formats, such as TIFF, and microscopy file formats, such as ND2. `Rainbow` can be easily used through a graphical user interference or command line interface by non-programmers. Importantly, `Rainbow` is not limited to ALI culture image analysis, and developers can extend the software’s existing pipeline to other use cases. For example, the optical flow model can be readily substituted with different models, as we utilised object orientated software design principles and data analyses and generated reports can be adjusted through interactive Jupyter Notebook documents, allowing for a flexible and versatile system. Some of `Rainbow`’s visualizations are shown in \autoref{fig1}.
+`Rainbow` is a free, open-source and cross-platform Python-based tool for Air-Liquid Interface (ALI) cell culture image analysis. `Rainbow` quantifies cell motion in image sequences using a recent state-of-the-art deep learning based optical flow model  [@jiangLearningEstimateHidden2021]. Optical flow is the apparent motion of individual pixels on an image plane across multiple frames [@turagaAdvancesVideoBasedHuman2010]. Our software then generates cell motion magnitude and direction variance-based metrics to provide quantitative insight into heterogeneous cell dynamics observed in the ALI culture images, in addition to commonly used metrics such as average cell speed. We also include a unified hue/saturation-based visualisation scheme for cell motion which offers high resolution as every coordinate on an image is utilised. `Rainbow` performs data analysis detailing cell dynamics by generating a HTML report containing output images, videos, publication ready figures, and p-values from input image sequences provided in standard image formats, such as TIFF, or microscopy file formats, such as ND2. `Rainbow` can be easily used through a graphical user interface or command line interface by non-programmers. Importantly, `Rainbow` is not limited to ALI culture image analysis, and developers can extend the software’s existing pipeline to other use cases. For example, the optical flow model can be readily substituted with different models, as we utilised object orientated software design principles, and the data analyses and reports generated can be adjusted through interactive Jupyter Notebook documents, allowing for a flexible and versatile system. Some of `Rainbow`’s visualizations are shown in \autoref{fig1}.
 
 # Statement of need
 
[…]
