### 2021-11-22 #3944 @AlphonsG
> @AlphonsG I am using the correct branch there, i.e. `dev` to generate the paper from?

Hello @Kevin-Mattheus-Moerman, nice to virtually meet you. I have merged `dev` with the main branch so please use the main branch to review the paper and software. Apologies for forgetting to merge.
