### 2022-01-15 #3944 @slee5777
> Hello,
>
> May I suggest Sarada Lee ([leesarada@gmail.com](mailto:leesarada@gmail.com)) and Sean Driver ([sean@assistedevolution.net](mailto:sean@assistedevolution.net)) who both have github accounts and experience in the field. Here is a recent publication of theirs:
>
> https://pubmed.ncbi.nlm.nih.gov/34413367/

Could you remove Sean and my personal email addresses?  We don't want them to make public.
