### 2022-01-17 #3944 @YuliyaLab
Hello @jmschrei ,
possible to get an estimate on the turn around time on the review? One of the co-authors here. Is this not the right avenue for our software paper? Can you please consider AlphonsG's suggestion above for the reviewer to speed things up?
Many thanks!
Yuliya
