### 2022-01-21 #4080 @whedon
[REVIEW]: Rainbow: Automated Air-Liquid Interface Cell Culture Analysis Using Deep Optical Flow
**Submitting author:** @AlphonsG (<a href="http://orcid.org/0000-0002-9607-2839">Alphons Gwatimba</a>)
**Repository:** https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0
