### 2022-01-21 AlphonsG/Rainbow-Optical-Flow-For-ALI#21 @Assistedevolution
Initial review comments and suggestions
Thank you for sharing this project. It’s great to see work being done in this area and development of open source tools for researchers.

My understanding is the aim of this project is primarily to provide a tool for high throughput analysis of ALI cultures, but also more broadly to provide a tool for the more generic problem of generating statistics for partially occluded cell tracking. A key focus is on quantification.
A driver of the work is to improve understanding of “airway cells” susceptibility to migration under different stimuli.
To achieve this, you have focused on quantification of cell motion, in particular cell motion magnitude and direction variance based metrics.
From the above you create a user friendly “no code” report with flow visualizations and statistics.

Some initial areas the pop out that could do with some clarity are:

Magnitude and direction variance-based metrics:
What is the frame of reference for variance of motion?
Is the analysis plotting instantaneous cell motion at a time T (frame of reference of the individual cell), or variance of motion and magnitude at a given pixel location (frame of reference the pixel)?

Related to the frame, Is the Hue/Saturation based visualization an interpolation, the instantaneous vector field of motion or individual cell motion by frame?
In your example, the Rainbow visualization does not seem to share common attributes with the raw image. Maybe make add a third picture of the two overlaid

Possibly worth stating that measurement of motility is a key focus and why
If motility is important, you should consider degree of heterogeneity, how magnitude and direction vary across the field of view. This is captured well in the vector plots. Would be interested to see some discussion of spatial variability of the vector field and how this might be captured and its quantitative interpretation in the pEMT / UJT regimes.

The work focuses on macro-optical flow but does not seem to look at cell shape and ratios of cells of different aspect ratios/ morphologies which have been addressed in the references. Possibly worth discussion about how the macro flow measurements might be upscaled to cell population statistics?

With respect to the reports: There probably needs to be more information on how the measured information can be used in a quantitative sense. This would suggest there is some baseline against which you are measuring normalcy. Visual indicators of flow often require further quantification methods before they can be used analytically. Could you elaborate, or is this beyond the scope of the current project?
Calculated element like fluid shear reflects directly on interfacial properties You may want to look at some of the MIT “video amplification for discovery of physical properties” work for some ideas there.

Some thought into how the individual Jupyter reports might be used by the researcher to derive experimental results, when dealing with multiple samples/cohorts. This may need addressing to make the system more directly usable without support?

Could you expand on the RAFT / GMA Training: was the GMA network retrained to your data or was the model trained on the Sintel dataset? Is the Sintel dataset benchmark considered appropriate for dynamically interacting bodies in large scale flow?

It may be useful to add details and comments on size limitation of input image for the neural net? GMA uses an attention based KQV model which can be computationally expensive to scale, which often limits the size of the input image.

Generally, it can be seen that a lot of thought has gone in to finding an appropriate technology to address motion capture analytics.
              Small gaps appear to be in:
                Possibly more background (for the new reader) on the problem being solved.
                Cleaner layout of Statement of need and summary
                Possibly spilt state of the field into separate section
                Comments on use / setup in lab, types of cameras size of image etc. FPS video format etc.
               Quantitative analysis and significance of flow information
	       Ability to extend beyond the single sample (how to compare multiple reports?)
	      Discussion of applicability of GMA network when applied to new data set (fine tuning)
	      More refences in the field of optical flow and citations for other opensource libs used
