### 2022-01-24 AlphonsG/Rainbow-Optical-Flow-For-ALI#21 @AlphonsG
Hello @Assistedevolution, thank you for your feedback. I believe all comments pertaining to the review are to be directed [here](https://github.com/openjournals/joss-reviews/issues/4080#event-5935625422), where there are other important items such as a reviewer checklist and so on. I'll close this issue and if you copy paste these comments there, the co-authors and I can go through them. Thanks.
