### 2022-01-24 #4080 @AlphonsG
Hello @jmschrei, just want to confirm if [these](https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI/issues/21) suggestions/questions should be there or in this issue page?
