### 2022-02-05 #4080 @Assistedevolution
Provided author feedback following review of paper and repository.
Feedback requested providing clarity around what was being measured and why, More thought suggested into end use of the information. Requested more background (for the new reader) on the problem being solved.

Flagged in repo fixable package versioning issues with installation under Anaconda, (Note Miniconda is recommended, however I feel its more likely that people will try this with Anaconda).
Waiting on resolution of issues
