### 2022-02-07 61f0164
Explain scripts folder
