+++ README.md
@@ -99,6 +99,10 @@ Rainbow uses a deep learning model called [GMA](https://arxiv.org/abs/2104.02409
 
 Rainbow can automatically generate an analysis report after computing the optical flow in an image series. A base report file that can be modified is provided [here](misc/notebooks/report.ipynb) as a Jupyter notebook. The path of a Jupyter notebook needs to specified in the config for automatic report generation (default provided).
 
+### Scripts
+
+The [scripts](scripts) folder contains python scripts to enable additional functionality such as the ability to combine reports from multiple experiments into one file for simpler viewing and comparisons.
+
 ## Examples <a name="examples"></a>
 
  Examples of some of the data generated by Rainbow can be seen below.
