### 2022-02-08 c0d79eb
Update paper
Based on reviewer suggestions
