+++ paper/paper.md
@@ -40,19 +40,19 @@ bibliography: paper.bib
 
 # Summary
 
-`Rainbow` is a free, open-source and cross-platform Python-based tool for Air-Liquid Interface (ALI) cell culture image analysis. `Rainbow` quantifies cell motion in image sequences using a recent state-of-the-art deep learning based optical flow model  [@jiangLearningEstimateHidden2021]. Optical flow is the apparent motion of individual pixels on an image plane across multiple frames [@turagaAdvancesVideoBasedHuman2010]. Our software then generates cell motion magnitude and direction variance-based metrics to provide quantitative insight into heterogeneous cell dynamics observed in the ALI culture images, in addition to commonly used metrics such as average cell speed. We also include a unified hue/saturation-based visualisation scheme for cell motion which offers high resolution as every coordinate on an image is utilised. `Rainbow` performs data analysis detailing cell dynamics by generating a HTML report containing output images, videos, publication ready figures, and p-values from input image sequences provided in standard image formats, such as TIFF, or microscopy file formats, such as ND2. `Rainbow` can be easily used through a graphical user interface or command line interface by non-programmers. Importantly, `Rainbow` is not limited to ALI culture image analysis, and developers can extend the software’s existing pipeline to other use cases. For example, the optical flow model can be readily substituted with different models, as we utilised object orientated software design principles, and the data analyses and reports generated can be adjusted through interactive Jupyter Notebook documents, allowing for a flexible and versatile system. Some of `Rainbow`’s visualizations are shown in \autoref{fig1}.
+`Rainbow` is a free, open-source and cross-platform Python-based tool for Air-Liquid Interface (ALI) cell culture image sequence analysis. `Rainbow` accepts input image sequences in standard image formats, such as TIFF, or microscopy file formats, such as ND2, and then computes the optical flow, that is: the apparent motion of individual pixels in an image [@turagaAdvancesVideoBasedHuman2010], across multiple frames using a deep learning based pretrained optical flow model [@jiangLearningEstimateHidden2021]. `Rainbow` then uses the pixel-level optical flow information and applies circular data analysis to calculate the average magnitude [0, ∞] (μm) and direction [0, 360) (°) of motion between adjacent frames to quantify cell motility. Additionally, the variance of the magnitude [0, ∞] (μm) and direction [0, ∞] (°) of motion between adjacent frames is calculated to quantitatively capture the degree of heterogeneity in cell motility. For each experiment, a CSV file with minimum, maximum, mean, standard deviation and variance values of the magnitude and direction of cell movement between adjacent frames in an image sequence is produced. Multiple CSV files from different experiments can be combined into one file that can be analysed for differences in cell motility across experiments. `Rainbow` also includes a high-resolution and easily readable unified hue/saturation-based visualisation scheme for the instantaneous vector field of motion between adjacent frames of an image sequence to qualitatively show cell motion. `Rainbow` can be used through a graphical user interface or command line interface and can generate a HTML report containing output images, videos, publication ready figures, and CSV files detailing cell dynamics (refer to Examples folder on GitHub) in one command. Importantly, our software is not limited to ALI culture image analysis, and developers can extend the software’s existing pipeline to other use cases. For example, the optical flow model can be readily substituted with different models, as we utilised the Factory Method creational software design pattern. The data analyses and report generated can be adjusted through interactive Jupyter Notebooks, allowing for a flexible and versatile system. Some of `Rainbow`’s visualizations are shown in \autoref{fig1}.
 
 # Statement of need
 
[…]
