+++ paper/paper.bib
@@ -207,3 +207,149 @@ Publisher: Nature Publishing Group},
 	keywords = {Computer Science - Computer Vision and Pattern Recognition},
 	file = {arXiv Fulltext PDF:C\:\\Users\\Alphons\\Zotero\\storage\\AAK5ZUTH\\Jiang et al. - 2021 - Learning to Estimate Hidden Motions with Global Mo.pdf:application/pdf;arXiv.org Snapshot:C\:\\Users\\Alphons\\Zotero\\storage\\PP5V48RM\\2104.html:text/html},
 }
+
+@article{harris2020array,
+	title = {Array programming with {NumPy}},
+	volume = {585},
+	url = {https://doi.org/10.1038/s41586-020-2649-2},
+	doi = {10.1038/s41586-020-2649-2},
+	number = {7825},
+	journal = {Nature},
+	author = {Harris, Charles R. and Millman, K. Jarrod and van der Walt, Stéfan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and van Kerkwijk, Marten H. and Brett, Matthew and Haldane, Allan and del Río, Jaime Fernández and Wiebe, Mark and Peterson, Pearu and Gérard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
+	month = sep,
+	year = {2020},
+	note = {Publisher: Springer Science and Business Media LLC},
+	pages = {357--362},
+}
+
+@article{Hunter:2007,
+	title = {Matplotlib: {A} {2D} graphics environment},
+	volume = {9},
+	doi = {10.1109/MCSE.2007.55},
+	abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.},
+	number = {3},
+	journal = {Computing in Science \& Engineering},
+	author = {Hunter, J. D.},
+	year = {2007},
+	note = {Publisher: IEEE COMPUTER SOC},
+	pages = {90--95},
+}
+
+@inproceedings{mckinney-proc-scipy-2010,
+	title = {Data {Structures} for {Statistical} {Computing} in {Python}},
+	doi = {10.25080/Majora-92bf1922-00a},
+	booktitle = {Proceedings of the 9th {Python} in {Science} {Conference}},
+	author = {{Wes McKinney}},
+	editor = {van der Walt, Stéfan and {Jarrod Millman}},
+	year = {2010},
+	pages = {56 -- 61},
+}
+
+@misc{tqdmref,
+	title = {tqdm: {A} fast, extensible progress bar for python and {CLI}},
+	url = {https://doi.org/10.5281/zenodo.5517697},
+	publisher = {Zenodo},
[…]
