+++ paper/paper.bib
@@ -325,7 +325,7 @@ tex.eid: A33},
 tex.adsnote: Provided by the SAO/NASA Astrophysics Data System
 tex.adsurl: https://ui.adsabs.harvard.edu/abs/2018AJ....156..123A
 tex.eid: 123},
-	keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, methods: data analysis, methods: miscellaneous, methods: statistical, reference systems},
+	keywords = {methods: data analysis, methods: miscellaneous, Astrophysics - Instrumentation and Methods for Astrophysics, methods: statistical, reference systems},
 	pages = {123},
 }
 
@@ -350,6 +350,31 @@ tex.eid: 123},
 	author = {van der Walt, Stéfan and Schönberger, Johannes L. and Nunez-Iglesias, Juan and Boulogne, François and Warner, Joshua D. and Yager, Neil and Gouillart, Emmanuelle and Yu, Tony and contributors, the scikit-image},
 	month = jun,
 	year = {2014},
-	keywords = {Education, Image processing, Open source, Python, Reproducible research, Scientific programming, Visualization},
+	keywords = {Visualization, Python, Image processing, Education, Open source, Reproducible research, Scientific programming},
 	pages = {e453},
+}
+
+@article{beauchemin1995computation,
+	title = {The computation of optical flow},
+	volume = {27},
+	number = {3},
+	journal = {ACM computing surveys (CSUR)},
+	author = {Beauchemin, Steven S. and Barron, John L.},
+	year = {1995},
+	note = {Publisher: ACM New York, NY, USA},
+	pages = {433--466},
+}
+
+@article{ZHAI2021107861,
+	title = {Optical flow and scene flow estimation: {A} survey},
+	volume = {114},
+	issn = {0031-3203},
+	url = {https://www.sciencedirect.com/science/article/pii/S0031320321000480},
+	doi = {https://doi.org/10.1016/j.patcog.2021.107861},
+	abstract = {Motion analysis is one of the most fundamental and challenging problems in the field of computer vision, which can be widely applied in many areas, such as autonomous driving, action recognition, scene understanding, and robotics. In general, the displacement field between subsequent frames can be divided into two types: optical flow and scene flow. The optical flow represents the pixel motion of adjacent frames. In contrast, the scene flow is a 3D motion field of the dynamic scene between two frames. Traditional approaches for the estimation of optical flow and scene flow usually leverage the variational technique, which can be solved as an energy minimization process. In recent years, deep learning has emerged as a powerful technique for learning feature representations directly from data. It has led to remarkable progress in the field of optical flow and scene flow estimation. In this paper, we provide a comprehensive survey of optical flow and scene flow estimation. First, we briefly review the pioneering approaches that use variational technique and then we delve in detail into the deep learning-based approaches. Furthermore, we present insightful observations on evaluation issues, specifically benchmark datasets, evaluation metrics, and state-of-the-art performance. Finally, we give the promising directions for future research. To the best of our knowledge, we are the first to review both optical flow and scene flow estimation, and the first to cover both traditional and deep learning-based approaches.},
+	journal = {Pattern Recognition},
+	author = {Zhai, Mingliang and Xiang, Xuezhi and Lv, Ning and Kong, Xiangdong},
+	year = {2021},
+	keywords = {Convolutional neural networks (CNNs), Deep learning, Motion analysis, Optical flow, Scene flow, Variational model},
+	pages = {107861},
 }
