### 2022-02-08 AlphonsG/Rainbow-Optical-Flow-For-ALI#21 @AlphonsG
Hello @Assistedevolution, thank you for your feedback and suggestions. Please find our responses below.

> Some initial areas the pop out that could do with some clarity are:
Magnitude and direction variance-based metrics:
What is the frame of reference for variance of motion? Is the analysis plotting instantaneous cell motion at a time T (frame of reference of the individual cell), or variance of motion and magnitude at a given pixel location (frame of reference the pixel)?
- The analysis is plotting the variance of motion and magnitude at a given pixel location (frame of reference the pixel). This has now been clarified in the paper.

> Related to the frame, Is the Hue/Saturation based visualization an interpolation, the instantaneous vector field of motion or individual cell motion by frame?
- The Hue/Saturation based visualization is the instantaneous vector field of motion. This has now been clarified in the paper.

> In your example, the Rainbow visualization does not seem to share common attributes with the raw image. Maybe make add a third picture of the two overlaid
- We assume that this comment refers to the figure in the paper. May the specific attributes not being shared please be clarified? The raw images (panels C and G) are included to help the reader visualize the original data that was used to generate Rainbow's visualizations i.e. the images in the other panels, such as the hue-saturation and quiver plots. For anything beyond that, including making comparisons between Rainbow’s visualization and the raw data, one will need to play the videos in the [examples folder](https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI/tree/dev2/examples/example_output/(tif)_191018_HNA-ALI_d14.nd2_-_191018_HNA-ALI_d14.nd2_(series_03)0000_etc) as now indicated in the paper.
> Possibly worth stating that measurement of motility is a key focus and why
- We have revised the Statement of need to highlight the involvement of cell motility in airway epithelial cell function and chronic respiratory disease development.

> If motility is important, you should consider degree of heterogeneity, how magnitude and direction vary across the field of view. This is captured well in the vector plots. Would be interested to see some discussion of spatial variability of the vector field and how this might be captured and its quantitative interpretation in the pEMT / UJT regimes.

- We have added information about the variance computation that Rainbow provides  as a measure of the degree of heterogeneity in cell motility, how magnitude and direction vary across the field of view. We are limited for space as per journal requirements, thus we avoid describing the pEMT / UJT regimes as part of this manuscript.

> The work focuses on macro-optical flow but does not seem to look at cell shape and ratios of cells of different aspect ratios/ morphologies which have been addressed in the references. Possibly worth discussion about how the macro flow measurements might be upscaled to cell population statistics?
>
- Due to scope and paper length limitations, we will skip a discussion on cell shape and ratios, but we have raised it as an area for further investigation and potential implementation in the future if there is a need. This [publication](https://www.nature.com/articles/nmat4357) provides a method for upscaling the macro flow measurements to cell population statistics which is compatible with our approach. We anticipate that obtaining the magnitude and direction as well as variances for these metrics will be the primary measures needed by researchers.

>With respect to the reports:
There probably needs to be more information on how the measured information can be used in a quantitative sense. This would suggest there is some baseline against which you are measuring normalcy. Visual indicators of flow often require further quantification methods before they can be used analytically. Could you elaborate, or is this beyond the scope of the current project?

- Currently, the cell migration field primarily focuses on measures of cell motility like magnitude of movement (speed) and direction of movement (angular distribution), as demonstrated by the cited references. Our open source software tool includes those and additional quantitative metrics (variance of movement) that provide novel insight to the user about cell biology. Relative comparisons between cohorts, or against an experimental ‘baseline’, are pertinent to individual biological use cases, and thus it is outside the scope of this project to provide in this software manuscript. Nevertheless, we have added a note that Rainbow generates CSV files with quantitative data that can be analysed, alongside the qualitative outputs such as images that are produced.

> Calculated element like fluid shear reflects directly on interfacial properties You may want to look at some of the MIT “video amplification for discovery of physical properties” work for some ideas there.
- We thank the reviewer for the suggestion. We will investigate this approach and potentially implement in future research.

> Some thought into how the individual Jupyter reports might be used by the researcher to derive experimental results, when dealing with multiple samples/cohorts. This may need addressing to make the system more directly usable without support?
- We have added the ability to combine reports from multiple experiments, as well as to combine quantitative analytical results from multiple experiments. We have provided an explanation in the paper. We will also develop and enhance this functionality over time.

> It may be useful to add details and comments on size limitation of input image for the neural net? GMA uses an attention based KQV model which can be computationally expensive to scale, which often limits the size of the input image.
- The authors of GMA do not place any limitations on the size of the input image (see: https://github.com/zacjiang/GMA/issues/6). Additionally, our example images (1280 x 1024 px) are a standard size for ALI imaging experiments and less than half the size of the images in [HD1K](http://hci-benchmark.iwr.uni-heidelberg.de/); one of the datasets used to train GMA by the original authors. Therefore, input image size should not be an issue.

>Generally, it can be seen that a lot of thought has gone in to finding an appropriate technology to address motion capture analytics.
Small gaps appear to be in:
Possibly more background (for the new reader) on the problem being solved.
- We have addressed this in earlier comments. Briefly, we have revised the Statement of need to indicate the role of airway epithelial cell motility in respiratory disease development, as well as the paucity of appropriate image analysis open-source software tools available for biological researchers in this area.

> Cleaner layout of Statement of need and summary
- We have revised the Statement of need and Summary.
> Possibly spilt state of the field into separate section
- We have revised this section.
> Comments on use / setup in lab, types of cameras size of image etc. FPS video format etc.
- We have added this information to the GitHub repository homepage.
> Quantitative analysis and significance of flow information
- This has been addressed in earlier comments.
> Ability to extend beyond the single sample (how to compare multiple reports?)
- This has been addressed in earlier comments.
> Could you expand on the RAFT / GMA Training: was the GMA network retrained to your data or was the model trained on the Sintel dataset? Is the Sintel dataset benchmark considered appropriate for dynamically interacting bodies in large scale flow? Discussion of applicability of GMA network when applied to new data set (fine tuning).
- We did not perform any finetuning. We have clarified the training aspect of the model in the paper. GMA should be applicable to a dataset with apparent motion, e.g. from cars or cells, since it processes pixels and not objects. Of course, the correctness of GMA’s results must be confirmed, which we did using the raw data and Rainbow’s using qualitative and quantitive outputs, which can be found in this GitHub repository and paper. Additionally, it is not unusual to use pretrained deep optical flow models without fine tuning on custom datasets due to the difficulty in generating ground truth data, hence why standard training datasets in the field are commonly synthetic/CGI based,, e.g. [Sintel](http://sintel.is.tue.mpg.de/).

> More refences in the field of optical flow and citations for other opensource libs used
- We have added more references to the paper.
