### 2022-02-08 AlphonsG/Rainbow-Optical-Flow-For-ALI#21 @Assistedevolution
Was thinking something like this for the image, Just makes visualising easier to flick back and forth between raw and processed data.

![Presentation1](https://user-images.githubusercontent.com/18102704/152933669-9e617ef9-b111-41ed-ad18-d1142335089d.gif)
