+++ requirements.txt
@@ -1,124 +1,130 @@
-anyio==3.3.4
-argon2-cffi==21.1.0
-astropy==5.0rc1
-atomicwrites==1.4.0
-attrs==21.2.0
-babel==2.9.1
-backcall==0.2.0
-bleach==4.1.0
-certifi==2021.10.8
-cffi==1.15.0
-charset-normalizer==2.0.7
-colorama==0.4.4
-colored==1.4.3
-cycler==0.11.0
-debugpy==1.5.1
-decorator==4.4.2
-defusedxml==0.7.1
-einops==0.3.2
-entrypoints==0.3
-fonttools==4.28.1
-gooey==1.0.8.1
-idna==3.3
-imageio==2.10.3
-imageio-ffmpeg==0.4.5
-importlib-resources==5.4.0
-imutils==0.5.4
-iniconfig==1.1.1
-ipykernel==6.5.0
-ipympl==0.8.2
-ipython==7.29.0
-ipython-genutils==0.2.0
-ipywidgets==8.0.0a6
-jedi==0.18.0
-jinja2==3.0.2
-json5==0.9.6
-jsonschema==4.2.1
-jupyter-client==7.0.6
-jupyter-core==4.9.1
-jupyter-server==1.11.2
-jupyterlab==4.0.0a14
-jupyterlab-pygments==0.1.2
-jupyterlab-server==2.8.2
-jupyterlab-widgets==2.0.0a3
-kiwisolver==1.3.2
-markupsafe==2.0.1
-matplotlib==3.5.0rc1
-matplotlib-inline==0.1.3
-mistune==0.8.4
-mkl-fft==1.3.1
-mkl-random==1.2.2
-mkl-service==2.4.0
-moviepy==2.0.0.dev2
-natsort==8.0.0
-nbclassic==0.3.4
-nbclient==0.5.5
-nbconvert==6.3.0b0
-nbformat==5.1.3
-nd2reader==3.3.0
-nest-asyncio==1.5.1
-networkx==2.6.3
-notebook==6.4.5
-numpy==1.21.2
-olefile==0.46
-opencv-python==4.5.4.58
-packaging==21.2
-pandas==1.3.4
-pandocfilters==1.5.0
-parso==0.8.2
-physt==0.5.1
-pickleshare==0.7.5
-Pillow==8.4.0
-pims==0.5
-pip==21.0.1
-pluggy==1.0.0
-proglog==0.1.9
-prometheus-client==0.12.0
-prompt-toolkit==3.0.22
-psutil==5.8.0
-py==1.11.0
-pycparser==2.21
-pyerfa==2.0.0.1
-pygments==2.10.0
-pygtrie==2.4.2
-pyparsing==2.4.7
-pyrsistent==0.18.0
-pytest==6.2.5
-python-dateutil==2.8.2
-pytz==2021.3
-pywavelets==1.2.0
-pywin32==302
-pywinpty==1.1.5
-pyyaml==6.0
-pyzmq==22.3.0
-requests==2.26.0
-scikit-image==0.18.3
-scipy==1.7.2
-send2trash==1.8.1b0
-setuptools==58.0.4
-setuptools-scm==6.3.2
-six==1.16.0
-slicerator==1.0.0
-sniffio==1.2.0
-terminado==0.12.1
-testpath==0.5.0
-tifffile==2021.11.2
-toml==0.10.2
-tomli==1.2.2
-torch==1.10.0
-torchaudio==0.10.0
-torchvision==0.11.1
-tornado==6.1
-tqdm==4.62.3
-traitlets==5.1.1
-typing-extensions==3.10.0.2
-urllib3==1.26.7
-wcwidth==0.2.5
[…]
