### 2022-02-08 f769f33
Update rainbow install command
