### 2022-02-08 f62843b
Fix ffmpeg log file silencing
