+++ requirements.txt
@@ -1,130 +1,127 @@
-Package             Version    Editable project location
-------------------- ---------- ----------------------------------------------------------------------------------------------------------------
-anyio               2.2.0
-argon2-cffi         20.1.0
-astropy             5.0.1
-async-generator     1.10
-atomicwrites        1.4.0
-attrs               21.4.0
-Babel               2.9.1
-backcall            0.2.0
-bleach              4.1.0
-brotlipy            0.7.0
-certifi             2021.10.8
-cffi                1.15.0
-charset-normalizer  2.0.4
-colorama            0.4.4
-colored             1.4.3
-cryptography        36.0.0
-cycler              0.11.0
-debugpy             1.5.1
-decorator           4.4.2
-defusedxml          0.7.1
-einops              0.4.0
-entrypoints         0.3
-fonttools           4.29.1
-Gooey               1.0.8.1
-idna                3.3
-imageio             2.15.0
-imageio-ffmpeg      0.4.5
-importlib-metadata  4.8.2
-imutils             0.5.4
-iniconfig           1.1.1
-ipykernel           6.4.1
-ipympl              0.8.7
-ipython             7.31.1
-ipython-genutils    0.2.0
-ipywidgets          7.6.5
-jedi                0.18.0
-Jinja2              3.0.2
-json5               0.9.6
-jsonschema          3.2.0
-jupyter-client      7.1.2
-jupyter-core        4.9.1
-jupyter-server      1.4.1
-jupyterlab          3.2.1
-jupyterlab-pygments 0.1.2
-jupyterlab-server   2.10.2
-jupyterlab-widgets  1.0.2
-kiwisolver          1.3.2
-MarkupSafe          2.0.1
-matplotlib          3.5.1
-matplotlib-inline   0.1.2
-mistune             0.8.4
-mkl-fft             1.3.1
-mkl-random          1.2.2
-mkl-service         2.4.0
-moviepy             2.0.0.dev2
-natsort             8.1.0
-nbclassic           0.2.6
-nbclient            0.5.3
-nbconvert           6.1.0
-nbformat            5.1.3
-nd2reader           3.3.0
-nest-asyncio        1.5.1
-networkx            2.6.3
-notebook            6.4.6
-numpy               1.21.5
-olefile             0.46
-opencv-python       4.5.5.62
-packaging           21.3
-pandas              1.4.0
-pandocfilters       1.5.0
-parso               0.8.3
-physt               0.5.3
-pickleshare         0.7.5
-Pillow              9.0.1
-PIMS                0.5
-pip                 22.0.3
-pluggy              1.0.0
-proglog             0.1.9
-prometheus-client   0.13.1
-prompt-toolkit      3.0.20
-psutil              5.9.0
-py                  1.11.0
-pycparser           2.21
-pyerfa              2.0.0.1
-Pygments            2.11.2
-pygtrie             2.4.2
-pyOpenSSL           22.0.0
-pyparsing           2.4.7
-pyrsistent          0.18.0
-PySocks             1.7.1
-pytest              7.0.0
-python-dateutil     2.8.2
[…]
