### 2022-02-09 #4080 @Assistedevolution
Successfully installed and ran code in the Dev2 branch.
Successfully Tested both in CLI and GUI mode.
Waiting on:
 -Some notes on documentation
 -Some notes on tests
 -Improvements in readability of paper layout
