+++ README.md
@@ -103,6 +103,10 @@ Rainbow can automatically generate an analysis report after computing the optica
 
 The [scripts](scripts) folder contains python scripts to enable additional functionality such as the ability to combine reports from multiple experiments into one file for simpler viewing and comparisons.
 
+### Automated Testing
+
+To perform and check the status of the automated tests locally, run the command `pytest` in the terminal from the root directory of this repository after cloning.
+
 ## Examples <a name="examples"></a>
 
  Examples of some of the data generated by Rainbow can be seen below.
