+++ requirements.txt
@@ -62,7 +62,7 @@ nd2reader==3.3.0
 nest-asyncio==1.5.1
 networkx==2.6.3
 notebook==6.4.6
-numpy==1.21.5
+numpy==1.22.2
 olefile==0.46
 opencv-python==4.5.5.62
 packaging==21.3
