+++ paper/paper.bib
@@ -118,7 +118,7 @@ Publisher: Nature Publishing Group},
 	copyright = {© 2018 John Wiley \& Sons Ltd},
 	issn = {1365-2222},
 	url = {http://onlinelibrary.wiley.com/doi/abs/10.1111/cea.13097},
-	doi = {https://doi.org/10.1111/cea.13097},
+	doi = {10.1111/cea.13097},
 	abstract = {Background Bronchial epithelial tight junctions (TJ) have been extensively assessed in healthy airway epithelium. However, no studies have yet assessed the effect of human rhinovirus (HRV) infection on the expression and resultant barrier function in epithelial tight junctions (TJ) in childhood asthma. Objectives To investigate the impact of HRV infection on airway epithelial TJ expression and barrier function in airway epithelial cells (AECs) of children with and without asthma. Furthermore, to test the hypothesis that barrier integrity and function is compromised to a greater extent by HRV in AECs from asthmatic children. Methods Primary AECs were obtained from children with and without asthma, differentiated into air-liquid interface (ALI) cultures and infected with rhinovirus. Expression of claudin-1, occludin and zonula occluden-1 (ZO-1) was assessed via qPCR, immunocytochemistry (ICC), in-cell western (ICW) and confocal microscopy. Barrier function was assessed by transepithelial electrical resistance (TER; RT) and permeability to fluorescent dextran. Results Basal TJ gene expression of claudin-1 and occludin was significantly upregulated in asthmatic children compared to non-asthmatics; however, no difference was seen with ZO-1. Interestingly, claudin-1, occludin and ZO-1 protein expression was significantly reduced in AEC of asthmatic children compared to non-asthmatic controls suggesting possible post-transcriptional inherent differences. HRV infection resulted in a transient dissociation of TJ and airway barrier integrity in non-asthmatic children. Although similar dissociation of TJ was observed in asthmatic children, a significant and sustained reduction in TJ expression concurrent with both a significant decrease in TER and an increase in permeability in asthmatic children was observed. Conclusion This study demonstrates novel intrinsic differences in TJ gene and protein expression between AEC of children with and without asthma. Furthermore, it correlates directly the relationship between HRV infection and the resultant dissociation of epithelial TJ that causes a continued altered barrier function in children with asthma.},
 	language = {en},
 	number = {5},
@@ -180,7 +180,7 @@ Publisher: Nature Publishing Group},
 	issn = {1439-0531},
 	shorttitle = {Air-liquid interface cell culture},
 	url = {http://onlinelibrary.wiley.com/doi/abs/10.1111/rda.13481},
-	doi = {https://doi.org/10.1111/rda.13481},
+	doi = {10.1111/rda.13481},
 	abstract = {The air-liquid interface (ALI) approach is primarily used to mimic respiratory tract epithelia in vitro. It is also known to support excellent differentiation of 3D multilayered skin models. To establish an ALI culture, epithelial cells are seeded into compartmentalized culture systems on porous filter supports or gel substrata. After an initial propagation period, the culture medium is removed from the apical side of the epithelium, exposing the cells to the surrounding air. Therefore, nutritive supply to the cells is warranted only by the basolateral cell pole. Under these conditions, the epithelial cells differentiate and regain full baso-apical polarity. Some types of epithelia even generate in vivo-like apical fluid or mucus. Interestingly, the ALI culture approach has also been shown to support morphological and functional differentiation of epithelial cells that are not normally exposed to ambient air in vivo. This review aims at giving a brief overview on the characteristics of ALI cultures in general and ALI models of female reproductive tract epithelia in particular. We discuss the applicability of ALI models for the investigation of the early embryonic microenvironment and for its implications in assisted reproductive technologies.},
 	language = {en},
 	number = {S3},
