### 2022-02-21 c3909f2
Add more automated testing detail in readme
