+++ requirements.txt
@@ -9,6 +9,7 @@ Babel==2.9.1
 backcall==0.2.0
 black==22.1.0
 bleach==4.1.0
+calver==2021.7.30
 certifi==2021.10.8
 cffi==1.15.0
 charset-normalizer==2.0.12
