### 2022-02-21 #4080 @tpeulen
@AlphonsG I am really sorry for the delay.

**Software:**
I already tried your software before accepting the review and had no issues. You could check all the boxed required for a publication in JOSS. I really liked your software. I only recommend you to additionally provide a conda environment yaml file (to facilitate the installation). Mixing conda and pip did not work very well for me.

**Manuscript:**
I can only suggest to spend more effort when writing a manuscripts. Frankly, the initial submission was not very well written and it seems that the reviewer @Assistedevolution spent more time for writing the review then the authors spend time when writing their manuscript.

**Overall:**
_Rainbow: Automated Air-Liquid Interface Cell Culture Analysis Using Deep Optical Flow_ (in its current revision) is of sufficient quality for publication in JOSS and I recommend publishing the manuscript.

@jmschrei Somehow I cannot edit the review checklist
