### 2022-02-22 f2070cd
Update build status badge
Now points to main instead of dev2 branch
