+++ setup.py
@@ -77,7 +77,7 @@ setup(
             'rainbow = rainbow.__main__:main'
         ]
     },
-    python_requires='>=3.6',
+    python_requires='>=3.8',
     cmdclass={
         'develop': PostDevelop,
         'install': PostInstall,
