### 2022-02-22 be1cde8
Remove gooey launcher workaround
Since no longer using conda.
