+++ CONTRIBUTING.md
@@ -31,4 +31,4 @@ Always write a clear log message for your commits, for example:
  ## Seek Support
  
  Please feel free to send an email to the following address for other matters:
- - alphonsg@protonmail.com
+ - 0go0vdp95@mozmail.com
