### 2022-03-02 #4080 @jmschrei
@tpeulen we recently switched the backend of our editorial bot (after the start of this review) so I can't re-add you to this review, but I've been told that you can write `@editorialbot generate my checklist`. Will that work for you?
