+++ paper/paper.md
@@ -16,10 +16,10 @@ authors:
   - name: Joseph Ho
     orcid: 0000-0002-9496-506X
     affiliation: "1, 2"
-  - name: Thomas Iosifidis^[co-second author]
+  - name: Thomas Iosifidis^[co-senior author]
     orcid: 0000-0001-8462-5865
     affiliation: "1, 3, 4"
-  - name: Yuliya V. Karpievitch^[co-second author]
+  - name: Yuliya V. Karpievitch^[co-senior author]
     orcid: 0000-0003-2586-3358
     affiliation: "1, 5"
 
@@ -30,7 +30,7 @@ affiliations:
     index: 2
   - name: Centre for Cell Therapy and Regenerative Medicine, School of Medicine, University of Western Australia, Nedlands, 6009, Western Australia, Australia
     index: 3
-  - name: School of Public Health, Curtin University, Bentley, 6102, Western Australia, Australia
+  - name: School of Population Health, Curtin University, Bentley, 6102, Western Australia, Australia
     index: 4
   - name: School of Molecular Sciences, University of Western Australia, Nedlands, 6009, Western Australia, Australia
     index: 5
@@ -40,19 +40,19 @@ bibliography: paper.bib
 
 # Summary
 
-`Rainbow` is a free, open-source and cross-platform Python-based tool for Air-Liquid Interface (ALI) cell culture time-lapse image analysis. `Rainbow` accepts time-lapse input images in standard image formats, such as TIFF, or microscopy file formats, such as ND2, and then computes the optical flow, that is: the apparent motion of individual pixels in an image [@turagaAdvancesVideoBasedHuman2010; @beauchemin1995computation; @ZHAI2021107861], across multiple frames using a deep learning based pretrained optical flow model [@jiangLearningEstimateHidden2021]. `Rainbow` then uses the pixel-level optical flow information and applies circular data analysis to calculate the average magnitude [0, ∞] (μm) and direction [0, 360) (°) of motion between adjacent frames to quantify cell motility. Additionally, the variance of the magnitude [0, ∞] (μm) and direction [0, ∞] (°) of motion between adjacent frames is calculated to quantitatively capture the degree of heterogeneity in cell motility.
+`Rainbow` is a free, open-source and cross-platform Python-based tool for Air-Liquid Interface (ALI) cell culture time-lapse image analysis. `Rainbow` accepts input time-lapse images in standard image formats, such as TIFF, or microscopy file formats, such as ND2, and then computes the optical flow; that is, the apparent motion of individual pixels in an image [@turagaAdvancesVideoBasedHuman2010; @beauchemin1995computation; @ZHAI2021107861], across multiple frames using a deep learning based pretrained optical flow model [@jiangLearningEstimateHidden2021]. `Rainbow` then applies circular data analysis to the pixel-level optical flow information to calculate the average magnitude [0, ∞] (μm) and direction [0, 360) (°) of motion between adjacent frames to quantify cell motility. Additionally, the variance of the magnitude [0, ∞] (μm) and direction [0, ∞] (°) of motion between adjacent frames is calculated to quantitatively capture the degree of heterogeneity in cell motility.
 
[…]
