### 2022-03-09 f41f9cc
Add more config documentation
