+++ requirements.txt
@@ -9,7 +9,7 @@ Babel==2.9.1
 backcall==0.2.0
 black==22.1.0
 bleach==4.1.0
-calver==2021.7.30
+bumpver==2022.1115
 certifi==2021.10.8
 cffi==1.15.0
 charset-normalizer==2.0.12
@@ -20,6 +20,7 @@ cycler==0.11.0
 debugpy==1.5.1
 decorator==4.4.2
 defusedxml==0.7.1
+docutils==0.18.1
 einops==0.4.0
 entrypoints==0.4
 executing==0.8.2
@@ -28,6 +29,7 @@ Gooey==1.0.8.1
 idna==3.3
 imageio==2.16.0
 imageio-ffmpeg==0.4.5
+importlib-metadata==4.11.1
 importlib-resources==5.4.0
 imutils==0.5.4
 iniconfig==1.1.1
@@ -47,7 +49,9 @@ jupyterlab==3.2.9
 jupyterlab-pygments==0.1.2
 jupyterlab-server==2.10.3
 jupyterlab-widgets==1.0.2
+keyring==23.5.0
 kiwisolver==1.3.2
+lexid==2021.1006
 MarkupSafe==2.1.0
 matplotlib==3.5.1
 matplotlib-inline==0.1.3
@@ -69,11 +73,13 @@ packaging==21.3
 pandas==1.4.1
 pandocfilters==1.5.0
 parso==0.8.3
+pathlib2==2.3.7.post1
 pathspec==0.9.0
 physt==0.5.3
 pickleshare==0.7.5
 Pillow==9.0.1
 PIMS==0.5
+pkginfo==1.8.2
 platformdirs==2.5.1
 pluggy==1.0.0
 proglog==0.1.9
@@ -86,17 +92,21 @@ pycparser==2.21
 pyerfa==2.0.0.1
 Pygments==2.11.2
 pygtrie==2.4.2
-pyparsing==3.0.7
+pyparsing==2.4.7
 pyrsistent==0.18.1
 pytest==7.0.1
 python-dateutil==2.8.2
 pytz==2021.3
 PyWavelets==1.2.0
 pywin32==303
+pywin32-ctypes==0.2.0
 pywinpty==1.1.6
 PyYAML==6.0
 pyzmq==22.3.0
+readme-renderer==32.0
 requests==2.27.1
+requests-toolbelt==0.9.1
+rfc3986==2.0.0
 scikit-image==0.19.2
 scipy==1.8.0
 Send2Trash==1.8.0
@@ -107,6 +117,7 @@ stack-data==0.2.0
 terminado==0.13.1
 testpath==0.5.0
 tifffile==2022.2.9
+toml==0.10.2
 tomli==2.0.1
 torch==1.10.2
 torchaudio==0.10.2
@@ -114,6 +125,7 @@ torchvision==0.11.3
 tornado==6.1
 tqdm==4.62.3
 traitlets==5.1.1
+twine==3.8.0
 typing_extensions==4.1.1
 urllib3==1.26.8
 wcwidth==0.2.5
