+++ bumpver.toml
@@ -1,0 +1,21 @@
+[bumpver]
+current_version = "2022.3.15"
+version_pattern = "YYYY.MM.INC0"
+commit_message = "bump version {old_version} -> {new_version}"
+commit = true
+tag = true
+push = false
+
+[bumpver.file_patterns]
+"setup.py" = [
+    "{version}",
+]
+"README.md" = [
+    "{version}",
+]
+"bumpver.toml" = [
+    'current_version = "{version}"',
+]
+"rainbow/__init__.py" = [
+    '__version__ = "{version}"',
+]
