### 2022-03-09 4c8d190
Update Readme for PyPi + cleanups
