+++ README.md
@@ -3,32 +3,33 @@
 ![](https://img.shields.io/badge/Rainbow-2022.3.15-blue)
 ![](https://img.shields.io/badge/platform-windows%20%7C%20linux%20%7C%20macOS-lightgrey)
 ![](https://img.shields.io/github/languages/top/AlphonsG/Rainbow-Optical-Flow-For-ALI)
-[![Python 3.6+](https://img.shields.io/badge/python-3.8+-blue.svg)](https://www.python.org/downloads/release/python-360/)
-[![Python package](https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI/actions/workflows/python-package.yaml/badge.svg?branch=dev2)](https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI/actions/workflows/python-package.yaml)
+[![Python 3.8+](https://img.shields.io/badge/python-3.8+-blue.svg)](https://www.python.org/downloads/)
+[![Python package](https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI/actions/workflows/python-package.yaml/badge.svg?branch=main)](https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI/actions/workflows/python-package.yaml)
 
-![](misc/images/banner_img.png?raw=true "Rainbow")
+<p style="text-align:center;">
+   <img src="https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI/raw/main/misc/images/banner_img.png" alt="centered image" />
+</p>
 
-Software for automated Air-Liquid Interface cell culture image analysis using deep optical flow. See [below](#software-paper) for more details.
+Software for automated Air-Liquid Interface cell culture image analysis using deep optical flow. See <a href="#software-paper">below</a> for more details.
 
 ## Table of contents
-1. [Installation](#installation)
-2. [Usage](#usage)
-3. [Additional Information](#additional-information)
-4. [Examples](#examples)
-6. [Community Guidelines](#community-guidelines)
-7. [License](#license)
-9. [Software Paper](#software-paper)
-10. [Our Team](#our-team)
-11. [Acknowledgements](#acknowledgements)
+1. <a href="#installation">Installation</a>
+2. <a href="#usage">Usage</a>
+3. <a href="#additional-information">Additional Information</a>
+4. <a href="#examples">Examples</a>
+5. <a href="#community-guidelines">Community Guidelines</a>
+6. <a href="#license">License</a>
+7. <a href="#software-paper">Software Paper</a>
+8. <a href="#our-team">Our Team</a>
+9. <a href="#acknowledgements">Acknowledgements</a>
 
-## Installation <a name="installation"></a>
+## Installation <a id="installation"></a>
 
 Rainbow can be installed on Linux, Windows & macOS and supports Python 3.8 and up. We recommend installing and running Rainbow within a [virtual environment](https://docs.python.org/3/tutorial/venv.html). Although it is not a requirement, we also recommend installing and running Rainbow on a GPU-enabled system to minimize processing times.
 
-1. Download and install [Python](https://www.python.org/downloads/) (Rainbow was tested using [Python version 3.8.10](https://www.python.org/downloads/release/python-3810/)), [Git](https://git-scm.com/downloads) and [Git LFS](https://git-lfs.github.com/).
+1. Download and install [Python](https://www.python.org/downloads/) (Rainbow was tested using [Python version 3.8.10](https://www.python.org/downloads/release/python-3810/)).
 
 2. Launch the terminal (*Linux* and *macOS* users) or command prompt (*Windows* users). The proceeding commands will be entered into the opened window.
-
 
[…]
