### 2022-03-09 4d8c7be
Alert user if PyTorch not installed
