+++ rainbow/__main__.py
@@ -36,6 +36,16 @@ def process_args():
 
 
 def main():
+    # check if pytorch is installed
+    try:
+        import torch
+    except ModuleNotFoundError:
+        msg = ('Rainbow cannot run because PyTorch is currently '
+               'not installed. Please install Pytorch by following the '
+               'instructions here: https://pytorch.org/get-started/locally/.')
+        print(msg)
+        exit(1)
+
     args = process_args()
     assert os.path.isdir(args.root_dir), ('Invalid root directory path '
                                           'provided.')
