### 2022-03-09 d482084
Update GMA class
Change checkpoint download routine
Change torch importing to only when methods are called
Fix gma_dir for pip package
