+++ rainbow/__init__.py
@@ -2,7 +2,7 @@
 #
 # This software is released under the MIT License.
 # https://opensource.org/licenses/MIT
-__version__ = "2022.3.15"
+__version__ = "2022.3.16"
 
 from rainbow.optical_flow.optical_flow import OPTICAL_FLOW_FILENAME
 
