+++ bumpver.toml
@@ -1,5 +1,5 @@
 [bumpver]
-current_version = "2022.3.15"
+current_version = "2022.3.16"
 version_pattern = "YYYY.MM.INC0"
 commit_message = "bump version {old_version} -> {new_version}"
 commit = true
