+++ pyproject.toml
@@ -1,2 +1,2 @@
 [build-system]
-requires = ["setuptools>=46", "wheel", "calver"]
+requires = ["setuptools>=46", "wheel"]
