+++ setup.py
@@ -47,7 +47,7 @@ setup(
     author='Alphons Gwatimba',
     author_email='0go0vdp95@mozmail.com',
     packages=find_packages(exclude=['tests*']),
-    version="2022.3.16",
+    version="2022.3.28",
     url='https://github.com/AlphonsG/Rainbow-Optical-Flow-For-ALI',
     license='MIT',
     classifiers=CLASSIFIERS,
