+++ paper/paper.bib
@@ -195,17 +195,18 @@ Publisher: Nature Publishing Group},
 }
 
 @article{jiangLearningEstimateHidden2021,
-	title = {Learning to {Estimate} {Hidden} {Motions} with {Global} {Motion} {Aggregation}},
-	url = {http://arxiv.org/abs/2104.02409},
-	abstract = {Occlusions pose a significant challenge to optical flow algorithms that rely on local evidences. We consider an occluded point to be one that is imaged in the first frame but not in the next, a slight overloading of the standard definition since it also includes points that move out-of-frame. Estimating the motion of these points is extremely difficult, particularly in the two-frame setting. Previous work relies on CNNs to learn occlusions, without much success, or requires multiple frames to reason about occlusions using temporal smoothness. In this paper, we argue that the occlusion problem can be better solved in the two-frame case by modelling image self-similarities. We introduce a global motion aggregation module, a transformer-based approach to find long-range dependencies between pixels in the first image, and perform global aggregation on the corresponding motion features. We demonstrate that the optical flow estimates in the occluded regions can be significantly improved without damaging the performance in non-occluded regions. This approach obtains new state-of-the-art results on the challenging Sintel dataset, improving the average end-point error by 13.6\% on Sintel Final and 13.7\% on Sintel Clean. At the time of submission, our method ranks first on these benchmarks among all published and unpublished approaches. Code is available at https://github.com/zacjiang/GMA},
-	urldate = {2021-10-28},
-	journal = {arXiv:2104.02409 [cs]},
-	author = {Jiang, Shihao and Campbell, Dylan and Lu, Yao and Li, Hongdong and Hartley, Richard},
-	month = jul,
-	year = {2021},
-	note = {arXiv: 2104.02409},
-	keywords = {Computer Science - Computer Vision and Pattern Recognition},
-	file = {arXiv Fulltext PDF:C\:\\Users\\Alphons\\Zotero\\storage\\AAK5ZUTH\\Jiang et al. - 2021 - Learning to Estimate Hidden Motions with Global Mo.pdf:application/pdf;arXiv.org Snapshot:C\:\\Users\\Alphons\\Zotero\\storage\\PP5V48RM\\2104.html:text/html},
+        author={Jiang, Shihao and Campbell, Dylan and Lu, Yao and Li, Hongdong and Hartley, Richard},  
+	booktitle={2021 IEEE/CVF International Conference on Computer Vision (ICCV)},   
+	title={Learning to Estimate Hidden Motions with Global Motion Aggregation},   
+	year={2021},  
+	volume={},  
+	number={},  
+	pages={9752-9761},  
[…]
