### 2022-03-17 #4080 @AlphonsG
> > Can confirm 10.1109/iccv48922.2021.00963 is valid a DOI for Learning to Estimate Hidden Motions with Global Motion Aggregation, the Arxiv reference given in the paper accurate..
>
> @AlphonsG, @YuliyaLab it seems one of your reviewers is going to the trouble of checking your missing DOIs for you! Can you please fix this now?

Hello, reference fixed now
