+++ paper/paper.md
@@ -48,7 +48,7 @@ For each experiment, a CSV file with the minimum, maximum, mean, standard deviat
 
 # Statement of need
 
-Assessment of airway epithelial cell (AEC) function in disease utilise in vitro models such as the ALI cell culture [@chenAirliquidInterfaceCell2019; @looiEffectsHumanRhinovirus2018; @martinovichConditionallyReprogrammedPrimary2017]. The integration of image analysis and ALI cell cultures has provided novel insights into cell dynamics, such as the recently identified unjammed-to-jammed transition of AEC characterized by changes in cell motility [@mitchelPrimaryAirwayEpithelial2020; @parkUnjammingCellShape2015]. In chronic respiratory diseases like asthma, increased cell motility and AEC unjamming have been linked to airway remodeling and disease development [@mitchelPrimaryAirwayEpithelial2020; @parkUnjammingCellShape2015]. However, the image analyses performed in these studies are limited. For example, handcrafted methods from the MATLAB Computer Vision Toolbox that compute optical flow have been used to extract cell motion information from ALI culture images [@mitchelPrimaryAirwayEpithelial2020]. This approach requires licenced software, and handcrafted optical flow estimation methods have been outperformed in terms of accuracy by deep learning methods [@savianOpticalFlowEstimation2020]. Furthermore, commonly used cell motion metrics, such as average cell speed, do not capture all unique aspects of cell motion, such as the heterogeneity of cell migration patterns across time. Cell motion is commonly visualized using vector fields, which are useful but bound by an inverse relationship between resolution and readability [@henkesDenseActiveMatter2020; @nnetuImpactJammingBoundaries2012; @osullivanIrradiationInducesEpithelial2020].
+Differentiated primary airway epithelial cells cultured at air-liquid interface (ALI) are commonly used to assess airway epithelial function in vitro in health and disease, such as asthma [@chenAirliquidInterfaceCell2019; @looiEffectsHumanRhinovirus2018; @martinovichConditionallyReprogrammedPrimary2017]. The integration of image analysis and ALI cell cultures has provided novel insights into cell dynamics, such as the recently identified unjammed-to-jammed transition of AEC characterized by changes in cell motility [@mitchelPrimaryAirwayEpithelial2020; @parkUnjammingCellShape2015]. In chronic respiratory diseases like asthma, increased cell motility and AEC unjamming have been linked to airway remodeling and disease development [@mitchelPrimaryAirwayEpithelial2020; @parkUnjammingCellShape2015]. However, the image analyses performed in these studies are limited. For example, handcrafted methods from the MATLAB Computer Vision Toolbox that compute optical flow have been used to extract cell motion information from ALI culture images [@mitchelPrimaryAirwayEpithelial2020]. This approach requires licenced software, and handcrafted optical flow estimation methods have been outperformed in terms of accuracy by deep learning methods [@savianOpticalFlowEstimation2020]. Furthermore, commonly used cell motion metrics, such as average cell speed, do not capture all unique aspects of cell motion, such as the heterogeneity of cell migration patterns across time. Cell motion is commonly visualized using vector fields, which are useful but bound by an inverse relationship between resolution and readability [@henkesDenseActiveMatter2020; @nnetuImpactJammingBoundaries2012; @osullivanIrradiationInducesEpithelial2020].
 
[…]
