### 2022-03-22 #4080 @AlphonsG
Thanks for the suggestions @danielskatz, I've updated the paper accordingly.
